+++ lib/simplepy.dart
@@ -1,6 +1,6 @@
 library;
 
-export 'src/parser.dart' show Parser;
+export 'src/parser.dart' show Parser, ParseError;
 export 'src/interpreter.dart' show Interpreter;
-export 'src/lexer.dart' show Lexer;
+export 'src/lexer.dart' show Lexer, LexerError;
 export 'src/ast_nodes.dart' show Stmt;